manejo de errores en subject service

+++ src/subjects/subjects.service.ts
@@ -20,14 +20,12 @@ export class SubjectsService {
       const { pupilDni, subjectName, qualification } = createSubjectDto;
       const subjectInstance = this.subjectFactory.createSubject(pupilDni, subjectName, qualification);
       const createdSubject = new this.subjectModel(subjectInstance);
-
+      createdSubject ? console.info("Se creo correctamente lña asignatura.") : console.error("Algo pasó al asignar los datos.");
       return await createdSubject.save();
 
     } catch (error) {
-
       throw new Error(`Ocurrió un error al crear la asignatura: \n
       ${error}`);
-
     }
   }
 
@@ -36,25 +34,63 @@ export class SubjectsService {
   }
 
   async findOne(pupilDni: string) {
-    return await this.subjectModel.findOne({ pupilDni }).exec();
+    try {
+
+      const subjectExist = await this.subjectModel.findOne({ pupilDni }).exec();
+      subjectExist ? console.info(`Se encontro el documento correspondiente al alumno con dni: ${pupilDni}`) : console.error(`Documento inexistente`);
+      return subjectExist;
+
+    } catch (error) {
+
+      throw new Error(`Ocurrió un error al buscar el documento solicitado: \n
+        ${error}`);
+
+    }
   }
 
   async update(id: string, updateSubjectDto: UpdateSubjectDto): Promise<Subject> {
-    const existingSubject = await this.subjectModel.findByIdAndUpdate(id).exec();
+    try {
 
-    if (!existingSubject) {
-      throw new Error(`The subjects with pupil dni: ${id} doesn't exists`);
+      const existingSubject = await this.subjectModel.findByIdAndUpdate(id).exec();
+
+      if (existingSubject)
+        throw new Error(`The subjects with pupil dni: ${id} doesn't exists`);
+      else
+        console.info("Se encontró el documento, procediendo con los cambios...")
+
+      const { pupilDni, subjectName, qualification } = updateSubjectDto;
+
+      if (existingSubject.pupilDni !== undefined && existingSubject.subjectName !== undefined
+        && existingSubject.qualification !== undefined) {
+        existingSubject.pupilDni = pupilDni;
+        existingSubject.subjectName = subjectName;
+        existingSubject.qualification = qualification;
+        console.info("Cambios realizados correctamente");
+
+        return await existingSubject.save();
+      } else {
+        throw new Error("No se realizaron los cambios solicitados.");
+      }
+
+    } catch (error) {
+
+      throw new Error(`Ocurrió un error al actualizar el documento: \n
+        ${error}`)
+
     }
-
-    const { pupilDni, subjectName, qualification } = updateSubjectDto;
-    existingSubject.pupilDni = pupilDni;
-    existingSubject.subjectName = subjectName;
-    existingSubject.qualification = qualification;
-
-    return await existingSubject.save();
   }
 
   async remove(id: string) {
-    return await this.subjectModel.findByIdAndDelete(id).exec();
+    try {
+      const subjectExists = await this.subjectModel.findById(id).exec();
+      if (subjectExists) {
+        await this.subjectModel.findByIdAndDelete(id).exec();
+        console.info("El documento se borró correctamente");
+      } else {
+        console.warn("No se encontró el documento solicitado para removerlo");
+      }
+    } catch (error) {
+      throw new Error("No se pudo borrar el documento");
+    }
   }
 }
